refactor: move serializing to its own file
